feat: get user organizations feature
